// src/types/enums/npcBattleState.ts
export enum NpcBattleState {
  NOT_STARTED = "NOT_STARTED",
  BATTLE = "BATTLE",
  PLAYER_TURN = "PLAYER_TURN",
  OPPONENT_TURN = "OPPONENT_TURN",
  PLAYER_ATTACKING = "PLAYER_ATTACKING",
  OPPONENT_ATTACKING = "OPPONENT_ATTACKING",
  PLAYER_POKEMON_FAINTED = "PLAYER_POKEMON_FAINTED",
  OPPONENT_POKEMON_FAINTED = "OPPONENT_POKEMON_FAINTED",
  PLAYER_SWITCHING_POKEMON = "PLAYER_SWITCHING_POKEMON",
  OPPONENT_SWITCHING_POKEMON = "OPPONENT_SWITCHING_POKEMON",
  PLAYER_VICTORY = "PLAYER_VICTORY",
  OPPONENT_VICTORY = "OPPONENT_VICTORY",
  BATTLE_ENDED = "BATTLE_ENDED",
  CANNOT_START = "CANNOT_START",
}